+++ react-marathon/src/routes/Game/index.js
@@ -1,0 +1,16 @@
+const GamePage = ({onChangePage}) => {
+    const handlerClick = () => {
+        onChangePage && onChangePage('home');
+    }
+
+    return (
+        <>
+            <p>Game Page!</p>
+            <button onClick={handlerClick}>
+                return
+            </button>
+        </>
+    );
+};
+
+export default GamePage;
